treewide: use built-in bool instead of np.bool
Use built-in bool type instead of np.bool, as np.bool "was a deprecated
alias of bool". AttributeError with the following message are thrown
when accessing np.bool on a newer version of NumPy:

AttributeError: module 'numpy' has no attribute 'bool'.

`np.bool` was a deprecated alias for the builtin `bool`. To avoid this
error in existing code, use `bool` by itself. Doing this will not modify
any behavior and is safe. If you specifically wanted the numpy scalar
type, use `np.bool_` here.
The aliases was originally deprecated in NumPy 1.20; for more details
and guidance see the original release note at:
https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations.

Did you mean: 'bool_'?